"some GPIO cleanup; updates #1171"
